Buttons were fixed. charge and Water pressure infos taken by the database to show device page.
The form submission process was repeated on each page refresh, affecting the information in the database. For this reason, the update method has been fixedand it has been adapted to the lights table in the database, allowing all ON/Off buttons to work.
Charge and water pressure information, which was written as a fixed value before, was taken from the database and made dynamic.